guix: build: propagated-inputs are now just inputs
Rationale: Propagated-inputs introduce problems like limiting the user
from installing a different version of the propagated-input
package. Since Emacsy is meant to be a library first, it should not
impose such limitations on the use of the library. The binaries should
still 'mostly' work.

This is a tradeoff: we'd rather accept that binary
fails to run than limit the use of library.