example: gtk-w-buffers now only processes events on keypress
This allows for emacsy to only process when an event occurs. For the
example this should be enough since we only have one event.

This fixes high CPU usages and should make key press updates as fast as
possible. It also fixes the need for an additional tick to catch
EMACSY_RAN_UNDEFINED_COMMAND_P during key press event.