build Emacsy libraries even when it's not possible to build examples
Syntax:
PKG_CHECK_MODULES(prefix,
  list-of-modules, action-if-found, action-if-not-found)

Add 'action-if-not-found' to set respective VARs to 'no', by default
action-if-not-found would be to throw an error and stop the build. But
we can still continue and build the libraries if not the examples.

This was found when trying to make a package for debian which had a
mismatching version of freeglut. <= 3.0